pyproject.toml
[project]
name = "customer-services"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk[eval]>=1.15",
    "google-cloud-aiplatform>=1.114.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "cloudpickle>=3.1.1",
    "uvicorn>=0.34.0",
    "fastapi>=0.118.0",
    "mcp[cli]>=1.15.0",
    "fastmcp>=2.12.4",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
